Added the ability to logout.
I'm not sure what the best practices are for this, but it works (look at
parts/loginlogout.jsp).  Fill free to change anything.

Also started on the ability to delete users.

Admin accounts should be available if you enter the admin key correctly
when creating an account (see util.Resources for the key).

When creating accounts you can no longer have two accounts with the same
email or same name.